test_infra/pyproject.toml
[tool.poetry]
name = "awswrangler - test infrastructure"
version = "2.12.1"
description = "CDK test infrastructure for AWS"
authors = ["Amazon Web Services"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = ">=3.6.2, <3.10"
"aws-cdk.core" = "^1.124.0"
"aws-cdk.aws-ec2" = "^1.124.0"
"aws-cdk.aws-glue" = "^1.124.0"
"aws-cdk.aws-iam" = "^1.124.0"
"aws-cdk.aws-kms" = "^1.124.0"
"aws-cdk.aws-logs" = "^1.124.0"
"aws-cdk.aws-s3" = "^1.124.0"
"aws-cdk.aws-redshift" = "^1.124.0"
"aws-cdk.aws-rds" = "^1.124.0"
"aws-cdk.aws-secretsmanager" = "^1.124.0"
"aws-cdk.aws-ssm" = "^1.124.0"
"aws-cdk.aws-opensearchservice" = "^1.124.0"
"aws-cdk.aws-lakeformation" = "^1.124.0"
